+++ libcaf_core/caf/function_view.hpp
@@ -52,20 +52,6 @@ private:
   std::tuple<Ts...>* storage_;
 };
 
-template <>
-class function_view_storage<unit_t> {
-public:
-  using type = function_view_storage;
-
-  explicit function_view_storage(unit_t&) {
-    // nop
-  }
-
-  void operator()() {
-    // nop
-  }
-};
-
 struct CAF_CORE_EXPORT function_view_storage_catch_all {
   message* storage_;
 
@@ -93,11 +79,6 @@ struct function_view_flattened_result {
 template <class T>
 struct function_view_flattened_result<std::tuple<T>> {
   using type = T;
-};
-
-template <>
-struct function_view_flattened_result<std::tuple<void>> {
-  using type = unit_t;
 };
 
 template <class T>
@@ -174,13 +155,24 @@ public:
     if (!impl_)
       return result_type{sec::bad_function_call};
     error err;
-    function_view_result<value_type> result;
-    self_->request(impl_, timeout, std::forward<Ts>(xs)...)
-      .receive([&](error& x) { err = std::move(x); },
-               typename function_view_storage<value_type>::type{result.value});
-    if (err)
-      return result_type{err};
-    return result_type{flatten(result.value)};
+    if constexpr (std::is_void_v<value_type>) {
+      self_->request(impl_, timeout, std::forward<Ts>(xs)...)
+        .receive([&](error& x) { err = std::move(x); }, [] {});
+      if (err)
+        return result_type{err};
+      else
+        return result_type{};
+    } else {
+      function_view_result<value_type> result;
+      self_->request(impl_, timeout, std::forward<Ts>(xs)...)
+        .receive([&](error& x) { err = std::move(x); },
+                 typename function_view_storage<value_type>::type{
+                   result.value});
+      if (err)
+        return result_type{err};
+      else
+        return result_type{flatten(result.value)};
+    }
   }
 
   void assign(type x) {